Removed AlignmentRecordRDDFunctions.
* Removed `AlignmentRecordRDDFunctions` and moved all functions into
  `AlignmentRecordRDD`. Moved test suite along with the main class move.
* Removed `countTag` functions and `PrintTags` CLI.
* Removed `JavaAlignmentRecordRDD`. The Java API just uses the
  `AlignmentRecordRDD` now. To do this, I moved the Java friendly save methods
  and JavaRDD to `GenomicRDD`.
