+++ adam-apis/src/main/scala/org/bdgenomics/adam/apis/java/JavaADAMContext.scala
@@ -20,8 +20,9 @@ package org.bdgenomics.adam.apis.java
 import org.apache.spark.SparkContext
 import org.apache.spark.api.java.JavaSparkContext
 import org.bdgenomics.adam.models.{ RecordGroupDictionary, SequenceDictionary }
-import org.bdgenomics.adam.rdd.ADAMContext
 import org.bdgenomics.adam.rdd.ADAMContext._
+import org.bdgenomics.adam.rdd.ADAMContext
+import org.bdgenomics.adam.rdd.read.AlignmentRecordRDD
 import org.bdgenomics.formats.avro._
 import scala.collection.JavaConversions._
 
@@ -60,10 +61,7 @@ class JavaADAMContext(val ac: ADAMContext) extends Serializable {
    * @param filePath Path to load the file from.
    * @return Returns a read RDD.
    */
-  def loadAlignments(filePath: java.lang.String): JavaAlignmentRecordRDD = {
-    val aRdd = ac.loadAlignments(filePath)
-    new JavaAlignmentRecordRDD(aRdd.rdd.toJavaRDD(),
-      aRdd.sequences,
-      aRdd.recordGroups)
+  def loadAlignments(filePath: java.lang.String): AlignmentRecordRDD = {
+    ac.loadAlignments(filePath)
   }
 }
