+++ adam-cli/src/main/scala/org/bdgenomics/adam/cli/ADAMMain.scala
@@ -60,7 +60,6 @@ object ADAMMain {
           PrintADAM,
           PrintGenes,
           FlagStat,
-          PrintTags,
           ListDict,
           AlleleCount,
           View
