+++ adam-cli/src/test/scala/org/bdgenomics/adam/cli/ADAM2FastqSuite.scala
@@ -55,11 +55,8 @@ class Adam2FastqSuite extends ADAMFunSuite {
     // Only looking at mapped reads
     // This is because Picard and ADAM disagree on setting negative strand on unmapped reads
     // Picard allows unmapped reads to set the negative strand flag and therefore reverse-complemented on output
-    val reads: RDD[AlignmentRecord] =
-      sc
-        .loadAlignments(readsFilepath)
-        .rdd
-        .filter(r => r.getReadMapped != null && r.getReadMapped)
+    val reads = sc.loadAlignments(readsFilepath)
+      .transform(_.filter(r => r.getReadMapped != null && r.getReadMapped))
 
     reads.saveAsFastq(outputFastqR1File, Some(outputFastqR2File), sort = true)
